requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
bcrypt==4.1.3
Pillow==10.3.0
pytesseract==0.3.10
python-magic==0.4.27
PyPDF2==3.0.1
openai>=1.0.0
azure-ai-documentintelligence==1.0.0b1
email-validator==2.1.1
boto3==1.34.103
PyMySQL==1.1.1
gunicorn==22.0.0
Werkzeug==3.0.2
